SyncController
SyncController now manages timer
