+++ standardnotes/AppDelegate.swift
@@ -16,8 +16,6 @@ import LocalAuthentication
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
     var window: UIWindow?
-    var syncTimer : Timer?
-    let syncTimeInterval : TimeInterval = 30
     weak var lockOutAlertVC: UIAlertController?
     
     static let sharedContext: NSManagedObjectContext = {
@@ -34,7 +32,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         Theme.Initialize()
         ItemManager.initializeSharedInstance(context: self.persistentContainer.viewContext)
         attemptFingerPrint()
-        startSyncTimer()
+        SyncController.sharedInstance.startSyncTimer()
         return true
     }
 
@@ -49,13 +47,13 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         if UserManager.sharedInstance.touchIDEnabled {
             navigateToAccountController(afterDelay: 0)
         }
-        stopSyncTimer()
+        SyncController.sharedInstance.stopSyncTimer()
     }
 
     func applicationWillEnterForeground(_ application: UIApplication) {
         // Called as part of the transition from the background to the active state; here you can undo many of the changes made on entering the background.
         attemptFingerPrint()
-        startSyncTimer()
+        SyncController.sharedInstance.startSyncTimer()
     }
 
     func applicationDidBecomeActive(_ application: UIApplication) {
@@ -128,17 +126,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         delay(afterDelay) {
             (UIApplication.shared.keyWindow?.rootViewController as? UITabBarController)?.selectedIndex = 0
         }
-    }
-    
-    func startSyncTimer() {
-        syncTimer?.invalidate()
-        syncTimer = Timer.scheduledTimer(withTimeInterval: syncTimeInterval, repeats: true, block: { (timer) in
-            ApiController.sharedInstance.sync(completion: { (error) in})
-        })
-    }
-    
-    func stopSyncTimer() {
-        syncTimer?.invalidate()
     }
     
     func attemptFingerPrint(){
